Ajustando Mapping e ajustando comando de update

+++ PaymentAPI.Domain/Features/Seller.cs
@@ -2,7 +2,7 @@
 {
     public class Seller
     {
-        public int SellerId { get; set; }
+        public long SellerId { get; set; }
         public string Name { get; set; }
         public int CPF { get; set; }
         public string Email { get; set; }
